# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "容器化Student",
        "title": "容器化Student",
        "termsOfService": "https://github.com/swaggo/gin-swagger",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/student": {
            "get": {
                "description": "根据姓名查询学生",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "学生"
                ],
                "summary": "根据姓名查询学生",
                "parameters": [
                    {
                        "type": "string",
                        "description": "姓名",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Student"
                        }
                    }
                }
            },
            "post": {
                "description": "新增学生信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "学生"
                ],
                "summary": "新增学生信息",
                "parameters": [
                    {
                        "description": "学生信息",
                        "name": "studentRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.Student"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/students": {
            "get": {
                "description": "查询所有学生",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "学生"
                ],
                "summary": "查询所有学生",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Student"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Student": {
            "type": "object",
            "properties": {
                "grade": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "student_id": {
                    "type": "string"
                }
            }
        },
        "request.Student": {
            "type": "object",
            "required": [
                "grade",
                "name",
                "student_id"
            ],
            "properties": {
                "grade": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "student_id": {
                    "type": "string"
                }
            }
        }
    }
}